requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
black==23.12.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
discord.py==2.3.2
ecdsa==0.18.0
frozenlist==1.4.1
idna==3.6
isort==5.13.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pyasn1==0.5.1
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
ruff==0.1.8
six==1.16.0
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
